+++ src/create.ts
@@ -3,6 +3,7 @@
 
 import { Mesh, MeshBuilder, TransformNode, Scene, Nullable, ActionManager, Tags, CreateGreasedLine, GreasedLineMeshBuilderOptions, Node, BoundingInfo, Observable, AbstractMesh } from '@babylonjs/core';
 import { createPlaneText } from './prefabs/Text/planeText';
+import { createContainer } from './prefabs/Misc/container';
 
 interface StringByFunc {
   [key: string]: Function;
@@ -20,42 +21,6 @@ function createGL(name: string, options: GreasedLineMeshBuilderOptions, scene: S
   return CreateGreasedLine(name, options, {}, scene);
 }
 
-interface containerOptions {
-  calculateBounds?: boolean,
-  exclude?: Nullable<(abstractMesh: AbstractMesh) => boolean> | undefined,
-  childObserver?: boolean,
-}
-
-function createContainer(name: string, options: containerOptions, scene: Scene){
-  let calculateBounds = options.calculateBounds || true;
-  let exclude = options.exclude || undefined;
-  let childObserver = options.childObserver || false; 
-
-  let container: Mesh = new Mesh(name, scene);
-
-  if (calculateBounds){
-    let { min, max } = container.getHierarchyBoundingVectors(true, exclude);
-    container.setBoundingInfo(new BoundingInfo(min, max));
-  }
-  
-  if (childObserver){
-    container.onAfterWorldMatrixUpdateObservable.add(() => {
-        setTimeout(() => {
-            if ((container as any).detectReentrancy) {
-                return;
-            }
-            (container as any).detectReentrancy = true;
-            let { min, max } = container.getHierarchyBoundingVectors(true, exclude); //triggers observable causing infinite loop
-            container.setBoundingInfo(new BoundingInfo(min, max));
-            setTimeout(() => {
-              (container as any).detectReentrancy = false;
-            });
-        });
-    })
-  }
-  
-  return container
-}
 
 const meshList: StringByFunc = {
   cot: createCOT,
